Extracted image shortcode to separate file
