+++ .eleventy.js
@@ -1,46 +1,10 @@
-// --- START, eleventy-img ---
-const Image = require("@11ty/eleventy-img");
-const path = require("path");
-
-async function imageShortcode(src, alt, sizes = "100vw", loading = "lazy") {
-  let metadata = await Image(src, {
-    widths: [300, 600, 1400],
-    formats: ["webp", "jpeg"],
-    outputDir: "_site/img",
-    filenameFormat: function (id, src, width, format) {
-      const extension = path.extname(src);
-      const name = path.basename(src, extension);
-      return `${name}-${width}w.${format}`;
-    },
-  });
-
-  let lowsrc = metadata.jpeg[0];
-  let highsrc = metadata.jpeg[metadata.jpeg.length - 1];
-
-  return `<picture>
-    ${Object.values(metadata).map(imageFormat => {
-      return `  <source type="${imageFormat[0].sourceType}" srcset="${imageFormat.map(entry => entry.srcset).join(", ")}" sizes="${sizes}">`;
-    }).join("\n")}
-      <img
-        src="${lowsrc.url}"
-        width="${highsrc.width}"
-        height="${highsrc.height}"
-        alt="${alt}"
-        loading="${loading}"
-        decoding="async">
-    </picture>`;
-
-}
-// --- END, eleventy-img ---
-
 const syntaxHighlight = require("@11ty/eleventy-plugin-syntaxhighlight");
 const pluginRss = require("@11ty/eleventy-plugin-rss");
 const directoryOutputPlugin = require("@11ty/eleventy-plugin-directory-output");
 
 module.exports = function (eleventyConfig) {
-  eleventyConfig.addNunjucksAsyncShortcode("image", imageShortcode);
 
-  // See directory output during build process
+  // Set directory output during build process
   eleventyConfig.setQuietMode(true);
   eleventyConfig.addPlugin(directoryOutputPlugin, {
     columns: {
@@ -55,16 +19,21 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.addPassthroughCopy("src/js");
   eleventyConfig.addPassthroughCopy("robots.txt");
 
-  // Shortcode for the current year
+  // Add shortcodes and filters
+
+  // shortcode: Eleventy image
+  eleventyConfig.addNunjucksAsyncShortcode("image", require("./src/_includes/shortcodes/image.js"));
+
+  // shortcode: current year
   eleventyConfig.addShortcode("year", () => `${new Date().getFullYear()}`);
 
-  // Add self-authored filter to generate reading time for a post
+  // filter: generate reading time for a post
   eleventyConfig.addFilter(
     "readingTime",
     require("./src/_includes/filters/readingtime.js")
   );
 
-  // Add filter to generate post date for various contexts
+  // filter: generate post date for various contexts
   eleventyConfig.addFilter(
     "formatPostDate",
     require("./src/_includes/filters/formatpostdate.js")
