+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "bobmonsour.com",
+  "name": "bm.com",
   "version": "1.0.0",
-  "description": "This is the personal website of Bob Monsour (built with 11ty).",
+  "description": "It's a work in progress. It's built with [eleventy](https://www.11ty.dev/)...something I'm still getting a handle on.",
   "main": ".eleventy.js",
   "scripts": {
     "watch:sass": "sass --watch src/sass/main-dev.scss:_site/css/main.css",
@@ -14,22 +14,22 @@
     "clean": "rimraf ./_site"
   },
   "keywords": [],
-  "author": "Bob Monsour",
+  "author": "",
   "license": "ISC",
   "devDependencies": {
     "@11ty/eleventy": "^1.0.1",
-    "@11ty/eleventy-img": "^1.1.0",
+    "@11ty/eleventy-img": "^2.0.1",
     "@11ty/eleventy-plugin-directory-output": "^1.0.1",
-    "@11ty/eleventy-plugin-rss": "^1.1.2",
-    "@11ty/eleventy-plugin-syntaxhighlight": "^4.0.0",
+    "@11ty/eleventy-plugin-rss": "^1.2.0",
+    "@11ty/eleventy-plugin-syntaxhighlight": "^4.1.0",
     "cross-env": "^7.0.3",
     "npm-run-all": "^4.1.5",
     "open-props": "^1.3.16",
     "path": "^0.12.7",
-    "postcss": "^8.4.12",
+    "postcss": "^8.4.14",
     "postcss-cli": "^9.1.0",
-    "postcss-jit-props": "^1.0.6",
+    "postcss-jit-props": "^1.0.7",
     "rimraf": "^3.0.2",
-    "sass": "^1.50.1"
+    "sass": "^1.53.0"
   }
 }
